Refactor choice classes to handle `this`
Each `Option` now holds a reference to the `Choice` instance that created it. This allows `toggle` to be called on an individual `Option` but still allows the parent instance to control overarching state (e.g. guaranteeing that only one option is toggled at a time)

Also refactored to use property initializers for the toggle method so that the classes are more easily composed without having to manually bind `this`